Add model valadation attributes and return http response messages to customer controller

+++ Controllers/CustomersController.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Net;
 using System.Threading.Tasks;
 using H_Plus_Sports.Models;
 using Microsoft.AspNetCore.Http;
@@ -22,43 +23,99 @@ namespace H_Plus_Sports.Controllers
         {
             _context = context;
         }
-
-        [HttpGet]
-        public IActionResult GetCustomer()
+        /// <summary>
+        /// Method correlated to GetCusotmer method, check to see if customer exists or not.
+        /// </summary>
+        /// <param name="id"></param>
+        /// <returns></returns>
+        private bool CustomerExists(int id)
         {
-            return new ObjectResult(_context.Customer);
+            return _context.Customer.Any(c => c.CustomerId == id);
         }
 
-        [HttpGet("{Id}", Name = "GetCustomer")]
+        [HttpGet]
+        [Produces(typeof(DbSet<Customer>))]
+        public IActionResult GetCustomer()
+        {
+            var results = new ObjectResult(_context.Customer)
+            {
+                StatusCode = (int)HttpStatusCode.OK
+            };
+
+            Request.HttpContext.Response.Headers.Add("X-Total-Count", _context.Customer.Count().ToString());
+
+            return results;
+        }
+
+        [HttpGet("{Id}")]
+        [Produces(typeof(Customer))]
         public async Task<IActionResult> GetCustomer([FromRoute] int id)
         {
+            if(!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var customer = await _context.Customer.SingleOrDefaultAsync(m => m.CustomerId == id);
+
+            if (customer == null)
+            {
+                return NotFound();
+            }
 
             return Ok(customer);
         }
 
         [HttpPost]
+        [Produces(typeof(Customer))]
         public async Task<IActionResult> PostCustomer([FromBody] Customer customer)
         {
+           if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             _context.Customer.Add(customer);
             await _context.SaveChangesAsync();
+
             return CreatedAtAction("getCustomer", new { id = customer.CustomerId }, customer);
         }
 
         [HttpPut("{id})")]
+        [Produces(typeof(Customer))]
         public async Task<IActionResult> PutCustomer([FromRoute] int id, [FromBody] Customer customer)
         {
+            if(!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
+            if (id != customer.CustomerId)
+            {
+                return BadRequest();
+            }
 
             _context.Entry(customer).State = EntityState.Modified;
-            await _context.SaveChangesAsync();
 
+            await _context.SaveChangesAsync();
             return Ok(customer);
         }
 
         [HttpDelete("{id})")]
+        [Produces(typeof(Customer))]
         public async Task<IActionResult> DeleteCustomer([FromRoute]int id)
         {
+            if(!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var customer = await _context.Customer.SingleOrDefaultAsync(m => m.CustomerId == id);
+            if (customer == null)
+            {
+                return NotFound();
+            }
+
             _context.Customer.Remove(customer);
             await _context.SaveChangesAsync();
